python/requirements.txt
requests
beautifulsoup4
redis
mypy
python-slugify
python-dotenv